Move random users establishment from test file to controllers

+++ part4/bloglistexpansion/controllers/blogs.js
@@ -48,7 +48,12 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-  // middle of this: await Blog.findById
+  /* middle of this: 
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if( blog.user.toString() === userid.toString() )
+  */ 
 
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
